PR comment - newline and renamed file
